* IntacctToggl subclass consolidated all intacct functionality to simply most use of Toggl * Reordered Toggl class methods, properties * renamed intacct methods * basic tests for IntacctToggl
